强制类型转换

+++ FiveChessServer.py
@@ -252,7 +252,7 @@ class ChessFactory(ServerFactory):
 
             else:
                 next_step = ai.GetAGoodMove(user["chessBoard"], user["level"])
-                a = next_step / 100
+                a = int(next_step / 100)
                 b = next_step - 100 * a
                 user["chessBoard"][b * 15 + a] = 0
